Rename heavydb_backend -> heavydb
