+++ rbc/heavydb/__init__.py
@@ -7,6 +7,7 @@ from .column_list import *  # noqa: F401, F403
 from .table_function_manager import *  # noqa: F401, F403
 from .text_encoding_dict import *  # noqa: F401, F403
 from .text_encoding_none import *  # noqa: F401, F403
+from .remoteheavydb import *  # noqa: F401, F403
 
 from . import mathimpl as math  # noqa: F401
 from . import npyimpl as np  # noqa: F401
