+++ rbc/heavydb/npyimpl.py
