+++ rbc/heavydb/table_function_manager.py
