+++ rbc/heavydb/text_encoding_dict.py
