+++ rbc/heavydb/text_encoding_none.py
